pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "crowdgit"
version = "0.0.1"
description = "Crowd.dev git integration for the Linux Foundation"
readme = "README.md"
authors = [
    { name = "Juan Reyero", email="juan@juanreyero.com" }
]
requires-python = ">=3.10"
license = { file="LICENSE" }
dependencies = [
    "requests >= 2.32.3",
    "tqdm",
    "fuzzywuzzy",
    "python-Levenshtein",
    "boto3 >= 1.35.18",
    "python-json-logger",
    "python-dotenv >= 1.0.1",
    "gitpython",
    "fastapi[standard]",
    "asyncio",
    "confluent_kafka",
    "openai >= 1.45.0",
    "prettytable >= 3.11.0",
    "psycopg2-binary >= 2.9.9",
    "python-slugify >= 8.0.4",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "jedi >= 0.18.1",
    "pylint >= 2.13.9",
    "pytest >= 7.0.0",
    "yapf >= 0.32.0"
]

[project.scripts]
crowd-git-ingest = "crowdgit.ingest:main"
crowd-git-bad-commits = "crowdgit.get_bad_commits:main"

[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=setup.py --ignore=build --ignore=doc --ignore=flymake"
pythonpath = [
  "."
]

[tool.black]
line-length = 99
target-version = ['py310']